Changing logic of rendering of floating but button, made its own component
